refactor: split `Table` types file into types dir w multiple files
