+++ src/Table/types/TableInstance.ts
@@ -1,0 +1,9 @@
+import type { TableKeysSchemaType } from "../../Schema/types/index.js";
+import type { Table } from "../Table.js";
+
+/**
+ * An instance of the {@link Table} class.
+ */
+export type TableInstance<TableKeysSchema extends TableKeysSchemaType> = InstanceType<
+  typeof Table<TableKeysSchema>
+>;
