+++ src/Table/types/index.ts
@@ -1,0 +1,5 @@
+export * from "./EnsureTableIsActiveParams.js";
+export * from "./TableConstructorParams.js";
+export * from "./TableCreateTableParameters.js";
+export * from "./TableInstance.js";
+export * from "./TableKeysAndIndexes.js";
